Even more cleanup and renaming
