+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/FrontendApplication.java
@@ -1,6 +1,5 @@
 package dev.totallyspies.spydle.frontend;
 
-import dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors.GameViewModel;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 
@@ -13,7 +12,7 @@ public class FrontendApplication {
         var ctx = new SpringApplicationBuilder(FrontendApplication.class).headless(false).run(args);
 
         EventQueue.invokeLater(() -> {
-            var gameView = ctx.getBean(GameViewModel.class);
+            var gameView = ctx.getBean(ViewManagerModel.class);
             gameView.setVisible(true);
         });
     }
