+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/ViewManagerModel.java
@@ -1,7 +1,7 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors;
+package dev.totallyspies.spydle.frontend;
 
 import dev.totallyspies.spydle.frontend.views.ListRoomsView;
-import dev.totallyspies.spydle.frontend.views.GameOverView;
+import dev.totallyspies.spydle.frontend.views.GameEndView;
 import dev.totallyspies.spydle.frontend.views.GameRoomView;
 import dev.totallyspies.spydle.frontend.views.WelcomeView;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +15,7 @@ import java.awt.*;
 Game View, which also acts as the window Frame
  */
 @Component
-public class GameViewModel extends JFrame {
+public class ViewManagerModel extends JFrame {
 
     // Define the CardLayout and panel container
     private final CardLayout cardLayout;
@@ -24,15 +24,15 @@ public class GameViewModel extends JFrame {
     private final WelcomeView welcomeView;
     private final GameRoomView gameRoomView;
     private final ListRoomsView listRoomsView;
-    private final GameOverView gameOverView;
+    private final GameEndView gameEndView;
 
     @Autowired // dependency injection
-    public GameViewModel(WelcomeView welcomeView, GameRoomView gameRoomView,
-                         ListRoomsView listRoomsView, GameOverView gameOverView) {
+    public ViewManagerModel(WelcomeView welcomeView, GameRoomView gameRoomView,
+                            ListRoomsView listRoomsView, GameEndView gameEndView) {
         this.welcomeView = welcomeView;
         this.gameRoomView = gameRoomView;
         this.listRoomsView = listRoomsView;
-        this.gameOverView = gameOverView;
+        this.gameEndView = gameEndView;
 
         // Set up the frame properties
         setTitle("SpyDle");
@@ -46,9 +46,9 @@ public class GameViewModel extends JFrame {
 
         // Add the different view panels to the card layout
         panelContainer.add(this.welcomeView, "WelcomeView");
-        panelContainer.add(this.listRoomsView, "AllRoomView");
+        panelContainer.add(this.listRoomsView, "ListRoomsView");
         panelContainer.add(this.gameRoomView, "GameRoomView");
-        panelContainer.add(this.gameOverView, "GameOverView");
+        panelContainer.add(this.gameEndView, "GameOverView");
 
         // Add the panel container to the frame
         add(panelContainer);
@@ -66,7 +66,7 @@ public class GameViewModel extends JFrame {
     public static void launchGameView(String[] args) {
         // Run the GameWindowFrame
         SwingUtilities.invokeLater(() -> {
-            GameViewModel frame = new GameViewModel(new WelcomeView(), new GameRoomView(), new ListRoomsView(), new GameOverView());
+            ViewManagerModel frame = new ViewManagerModel(new WelcomeView(), new GameRoomView(), new ListRoomsView(), new GameEndView());
             frame.setVisible(true);
         });
     }
