+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adapters/game_end/GameEndViewController.java
@@ -1,18 +1,18 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.all_room_adaptors;
+package dev.totallyspies.spydle.frontend.interface_adapters.game_end;
 
-import dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors.SwitchViewEvent;
+import dev.totallyspies.spydle.frontend.SwitchViewEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ListRoomsViewController {
+public class GameEndViewController {
 
     @Autowired
     private ApplicationEventPublisher publisher;
 
     /*
-    Method called when View All Rooms Button is Pressed
+    Method called when back Button is Pressed
      */
     public void openWelcomeView() {
         publisher.publishEvent(new SwitchViewEvent(this, "WelcomeView"));
