+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adapters/game_end/GameEndViewModel.java
@@ -1,0 +1,8 @@
+package dev.totallyspies.spydle.frontend.interface_adapters.game_end;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class GameEndViewModel {
+
+}
