+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adapters/game_room/GameRoomViewController.java
@@ -1,12 +1,12 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.welcome_screen_adaptors;
+package dev.totallyspies.spydle.frontend.interface_adapters.game_room;
 
-import dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors.SwitchViewEvent;
+import dev.totallyspies.spydle.frontend.SwitchViewEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
-public class WelcomeViewController {
+public class GameRoomViewController {
 
     @Autowired
     private ApplicationEventPublisher publisher;
@@ -14,8 +14,8 @@ public class WelcomeViewController {
     /*
     Method called when View All Rooms Button is Pressed
      */
-    public void openListRoomsView() {
-        publisher.publishEvent(new SwitchViewEvent(this, "ListRoomsView"));
+    public void openWelcomeView() {
+        publisher.publishEvent(new SwitchViewEvent(this, "WelcomeView"));
     }
 
 }
