+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adapters/list_rooms/ListRoomsViewController.java
@@ -1,12 +1,12 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.game_over_adaptors;
+package dev.totallyspies.spydle.frontend.interface_adapters.list_rooms;
 
-import dev.totallyspies.spydle.frontend.interface_adaptors.game_view_adaptors.SwitchViewEvent;
+import dev.totallyspies.spydle.frontend.SwitchViewEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 
 @Component
-public class GameOverViewController {
+public class ListRoomsViewController {
 
     @Autowired
     private ApplicationEventPublisher publisher;
