+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adapters/welcome/WelcomeViewPresenter.java
@@ -1,4 +1,4 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.welcome_screen_adaptors;
+package dev.totallyspies.spydle.frontend.interface_adapters.welcome;
 
 import org.springframework.stereotype.Component;
