--- frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adaptors/game_over_adaptors/GameOverViewModel.java
@@ -1,8 +1,0 @@
-package dev.totallyspies.spydle.frontend.interface_adaptors.game_over_adaptors;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class GameOverViewModel {
-
-}
